Fix indexing for JuMPArrays where upperbound is not a compile-time constant
The previous code I just committed was missing an ``esc``, which doesn't
play nicely with us evaling the code at toplevel. Instead, we can reuse
the offsets computed by taking advantage of the requirement that the
lowerbound is a compile-time constant. The code is also now shorter and
prettier, in addition to being correct. It's also significantly faster
than the release-0.9 iteration code (.945523066 s vs 15.951185824 s for
iterating over a 2000x1999 JuMPArray).