Querying after modifying is undefined behaviour
I do believe the change is self-explanatory. The only issue may be that it is too much text for a smal detail. However, the user needs the documentation to know what is undefined behavior, as if they test it programmatically they can be led to believe it works some way and not
another.