Querying after modifying is undefined behaviour (#2076)
* Querying after modifying is undefined behaviour

I do believe the change is self-explanatory. The only issue may be that it is too much text for a smal detail. However, the user needs the documentation to know what is undefined behavior, as if they test it programmatically they can be led to believe it works some way and not
another.

* Remove the example and use admonition box

I removed the example and reduced the test to a paragraph inside an admonition box following the syntax here (https://docs.julialang.org/en/v1/stdlib/Markdown/index.html#Admonitions-1). The box does not show in github preview, is this right?

* Fixed typo and removed exempli gratia

Just did what was asked by odow.

* Standardize spelling