Add lp_sensitivity_report
This functionality is intended as a replacement for the previous
lp_xxx_perturbation_range code. The main criticism of the previous
implementation was that it created the basis matrix every function
call, rather than doing it once and caching the result.

Instead of adding complexity by having a field cache the basis, this
PR computes the perturbation for every variable and constraint up-front.