Add a feasibility checker. (#2466)
This is a significant item for JuMP 1.0, but there is debate on how
to define and implement distances for different sets.

See https://github.com/matbesancon/MathOptSetDistances.jl for a WIP
package exploring the options.

This PR also only addresses primal feasibility. See
https://github.com/joaquimg/FeasibilityOptInterface.jl/pull/1
for a WIP attempt at dual feasibility as well.

In future, the _distance_to_set definitions will be merged into MOI.

This PR is more about defining the JuMP-level interface so we can
change internal details without breaking compatibility of JuMP 1.0.