WIP: remove zero() and one() for variables and expressions
See issue #1151 for background.

The problem is that calls like zero(AffExpr) lead to hard-to-diagnose
bugs.

However, much of Julia's LinearAlgebra routines assume that zero and
one are defined for the element types of arrays. Thus, if we do remove
these methods, we're likely to break a lot of user-code.