WIP: begin refatoring of nlp.jl
The nonlinear code in JuMP is spread across a range of files and submodules.
This is the first step towards documenting and refactoring (without breaking
changes) the NLP code in JuMP.

My goal is to move the NLPEvaluator code into _Derivatives and the
nonlinear expression graph stuff from _Derivatives into JuMP, so that we
separate the AD backend from the expression datastructure. That should
then let us swap in SymbolicAD without too much trouble.