Account for diagonal quadratic coefficients
Note that for diagonal quadratic coefficients of the form a*x[i]^2, the second derivative with respect to x[i] is 2*a. When adding the quadratic terms, we should check to see if the term is diagonal and compute the derivative accordingly. The reason we didn't notice this issue in the example provided here is that the optimal dual variable for the only quadratic constraint is 0, so this constraint doesn't show up in the optimal Hessian.