Refactor runtests.jl to use a single run_tests function
This PR continues the refactoring I have been doing of JuMP's tests.

 * Tests are now contained in a module which lives in test/test_xxx.jl
   file
 * Within a module, tests are functions beginning with test_
 * JuMPExtensions are tested by functions with the prefix test_extension_
   and take two optional positional argumennts for the model type and
   variable type
 * Test individual files via the command line \`julia runtests.jl [\<file\> [\<file\> ...]]\`

This design is flexible, and separates the runninng of the tests from their
definition. This should let us swap in other drivers instead of the one
currently in runtests.jl.

For example, we might want to experiment with:

 * Running all test functions within a module in parallel
 * Running each test file in parallel
 * Providing a more flexible command line interface

If we get something nice, we could package it into Pukeko.jl for wider use
in the Julia ecosystem.