Update variables.md explanation of parameters
I was confused by this example at multiple levels. I didn't understand what "Parameters set" means, is there a whole set of parameters somewhere? I didn't understand `Parameter(i)` because `i` looks like an index, not a value. I didn't get how there can be `x in SecondOrderCone` and `x in Parameter(i)`, same syntax but seemingly different concepts. The values are eventually revealed later on in example, but even then I was confused.

This change is to try to be more explicit about what's going on, in case others are as obtuse as I am. I hope something like this will be clearer.